+++ presentation/src/main/java/ru/vafeen/presentation/stopwatch/StopwatchDataState.kt
@@ -1,4 +1,4 @@
-package ru.vafeen.presentation.stop_watch
+package ru.vafeen.presentation.stopwatch
 
 import ru.vafeen.domain.domain_models.Stopwatch
