+++ presentation/src/main/java/ru/vafeen/presentation/stopwatch/StopwatchDataViewModel.kt
@@ -1,4 +1,4 @@
-package ru.vafeen.presentation.stop_watch
+package ru.vafeen.presentation.stopwatch
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
